Correct i d

+++ backend/src/projects/estimates/estimates.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  NotFoundException,
   Param,
   Patch,
   Post,
@@ -11,31 +12,39 @@ import {
 } from '@nestjs/common';
 import { EstimatesService } from './estimates.service';
 import { Types } from 'mongoose';
-import { ProjectGuard } from '../project/project.guard';
 import { EstimateDto } from './estimate.dto';
+import { Helpers } from '../positions/helpers';
+import { ErrorsApp } from 'src/common/errors';
+import { ProjectLargeGuard } from '../project/project.large.guard';
 
 @Controller('estimates')
 export class EstimatesController {
   constructor(private readonly estimatesService: EstimatesService) {}
   @Post(':projectId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLargeGuard)
   async create(
     @Body() dto: EstimateDto,
     @Param('projectId') projectId: string,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.estimatesService.createEstimate(dto, objectId);
   }
 
   @Patch('/:projectId/:estimateId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLargeGuard)
   async update(
     @Body() dto: EstimateDto,
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     return await this.estimatesService.updateEstimated(
@@ -46,11 +55,14 @@ export class EstimatesController {
   }
   @Delete('/:projectId/:estimateId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLargeGuard)
   async remove(
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     return await this.estimatesService.removeEstimate(
