+++ backend/src/projects/low-estimate/low-estimate/low.estimate.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  NotFoundException,
   Param,
   Patch,
   Post,
@@ -10,9 +11,11 @@ import {
   ValidationPipe,
 } from '@nestjs/common';
 import { LowEstimateService } from './low.estimate.service';
-import { ProjectGuard } from 'src/projects/project/project.guard';
 import { Types } from 'mongoose';
 import { EstimateDto } from 'src/projects/estimates/estimate.dto';
+import { Helpers } from 'src/projects/positions/helpers';
+import { ErrorsApp } from 'src/common/errors';
+import { ProjectLowGuard } from 'src/projects/project/project.low.guard';
 
 @Controller('low/estimate')
 export class LowEstimateController {
@@ -20,23 +23,30 @@ export class LowEstimateController {
 
   @Post(':projectId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLowGuard)
   async create(
     @Body() dto: EstimateDto,
     @Param('projectId') projectId: string,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.lowEstimateService.createLowEstimate(dto, objectId);
   }
 
   @Patch('/:projectId/:estimateId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLowGuard)
   async update(
     @Body() dto: EstimateDto,
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
+
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     return await this.lowEstimateService.updateEstimated(
@@ -48,11 +58,14 @@ export class LowEstimateController {
 
   @Delete('/:projectId/:estimateId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLowGuard)
   async remove(
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     return await this.lowEstimateService.removeEstimateLow(
