+++ backend/src/projects/low-estimate/low-position/low.position.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  NotFoundException,
   Param,
   Patch,
   Post,
@@ -10,9 +11,11 @@ import {
   ValidationPipe,
 } from '@nestjs/common';
 import { LowPositionService } from './low.position.service';
-import { ProjectGuard } from 'src/projects/project/project.guard';
 import { CreatePositionDto } from 'src/projects/positions/position-dto/position.create.dto';
 import { Types } from 'mongoose';
+import { Helpers } from 'src/projects/positions/helpers';
+import { ErrorsApp } from 'src/common/errors';
+import { ProjectLowGuard } from 'src/projects/project/project.low.guard';
 
 @Controller('low/position')
 export class LowPositionController {
@@ -20,12 +23,15 @@ export class LowPositionController {
 
   @Post('/:projectId/:estimateId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLowGuard)
   async create(
     @Body() dto: CreatePositionDto,
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     return await this.lowPositionService.createPosition(
@@ -37,13 +43,16 @@ export class LowPositionController {
 
   @Patch(':projectId/:estimateId/:positionId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLowGuard)
   async update(
     @Body() dto: CreatePositionDto,
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
     @Param('positionId') positionId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     await this.lowPositionService.updatePosition(
@@ -55,12 +64,15 @@ export class LowPositionController {
   }
 
   @Delete(':projectId/:estimateId/:positionId')
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLowGuard)
   async remove(
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
     @Param('positionId') positionId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     return await this.lowPositionService.removePosition(
