+++ backend/src/projects/low-project-price/low.project.price.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  NotFoundException,
   Param,
   Patch,
   Post,
@@ -10,9 +11,11 @@ import {
   ValidationPipe,
 } from '@nestjs/common';
 import { LowProjectPriceService } from './low.project.price.service';
-import { ProjectGuard } from '../project/project.guard';
 import { PricesDto } from 'src/prices/price.dto';
 import { Types } from 'mongoose';
+import { Helpers } from '../positions/helpers';
+import { ErrorsApp } from 'src/common/errors';
+import { ProjectLowGuard } from '../project/project.low.guard';
 
 @Controller('low/project/price')
 export class LowProjectPriceController {
@@ -22,20 +25,26 @@ export class LowProjectPriceController {
 
   @Post('/:projectId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLowGuard)
   async create(@Body() dto: PricesDto, @Param('projectId') projectId: string) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.lowProjectPriceService.createPrice(dto, objectId);
   }
 
   @Patch('/:projectId/:priceId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLowGuard)
   async update(
     @Body() dto: PricesDto,
     @Param('projectId') projectId: string,
     @Param('priceId') priceId: string,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.lowProjectPriceService.updatePrice(
       dto,
@@ -45,11 +54,14 @@ export class LowProjectPriceController {
   }
 
   @Delete('/:projectId/:priceId')
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLowGuard)
   async remove(
     @Param('projectId') projectId: string,
     @Param('priceId') priceId: string,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.lowProjectPriceService.removePrice(objectId, priceId);
   }
