+++ backend/src/projects/positions/positions.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  NotFoundException,
   Param,
   Patch,
   Post,
@@ -10,9 +11,11 @@ import {
   ValidationPipe,
 } from '@nestjs/common';
 import { PositionsService } from './positions.service';
-import { ProjectGuard } from '../project/project.guard';
 import { CreatePositionDto } from './position-dto/position.create.dto';
 import { Types } from 'mongoose';
+import { ErrorsApp } from 'src/common/errors';
+import { Helpers } from './helpers';
+import { ProjectLargeGuard } from '../project/project.large.guard';
 
 @Controller('positions')
 export class PositionsController {
@@ -20,12 +23,15 @@ export class PositionsController {
 
   @Post('/:projectId/:estimateId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLargeGuard)
   async create(
     @Body() dto: CreatePositionDto,
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     return await this.positionsService.createPosition(
@@ -37,13 +43,16 @@ export class PositionsController {
 
   @Patch(':projectId/:estimateId/:positionId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLargeGuard)
   async update(
     @Body() dto: CreatePositionDto,
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
     @Param('positionId') positionId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     await this.positionsService.updatePosition(
@@ -55,12 +64,15 @@ export class PositionsController {
   }
 
   @Delete(':projectId/:estimateId/:positionId')
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLargeGuard)
   async remove(
     @Param('projectId') projectId: string,
     @Param('estimateId') estimateId: string,
     @Param('positionId') positionId: string,
   ) {
+    if (!Helpers.checkId(projectId) || !Helpers.checkId(estimateId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectProjectId = new Types.ObjectId(projectId);
     const objectEstimatedId = new Types.ObjectId(estimateId);
     return await this.positionsService.removePosition(
