+++ backend/src/projects/project-prices/project.prices.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  NotFoundException,
   Param,
   Patch,
   Post,
@@ -12,7 +13,9 @@ import {
 import { ProjectPricesService } from './project.prices.service';
 import { PricesDto } from 'src/prices/price.dto';
 import { Types } from 'mongoose';
-import { ProjectGuard } from '../project/project.guard';
+import { Helpers } from '../positions/helpers';
+import { ErrorsApp } from 'src/common/errors';
+import { ProjectLargeGuard } from '../project/project.large.guard';
 
 @Controller('project/prices')
 export class ProjectPricesController {
@@ -20,30 +23,39 @@ export class ProjectPricesController {
 
   @Post('/:projectId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLargeGuard)
   async create(@Body() dto: PricesDto, @Param('projectId') projectId: string) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.projectPricesService.createPrice(dto, objectId);
   }
 
   @Patch('/:projectId/:priceId')
   @UsePipes(new ValidationPipe())
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLargeGuard)
   async update(
     @Body() dto: PricesDto,
     @Param('projectId') projectId: string,
     @Param('priceId') priceId: string,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.projectPricesService.updatePrice(dto, objectId, priceId);
   }
 
   @Delete('/:projectId/:priceId')
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectLargeGuard)
   async remove(
     @Param('projectId') projectId: string,
     @Param('priceId') priceId: string,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.projectPricesService.removePrice(objectId, priceId);
   }
