+++ backend/src/projects/setting-project/setting.project.controller.ts
@@ -1,6 +1,7 @@
 import {
   Body,
   Controller,
+  NotFoundException,
   Param,
   Patch,
   Post,
@@ -12,19 +13,24 @@ import { AddAllowDto } from './dto/add.allow.dto';
 import { RequestWithUser } from 'src/interfaces/requestWithUser';
 import { Types } from 'mongoose';
 import { DeleteAllowDto } from './dto/delete.dto';
-import { ProjectGuard } from '../project/project.guard';
 import { DiscountDto } from './dto/discount.dto';
+import { Helpers } from '../positions/helpers';
+import { ErrorsApp } from 'src/common/errors';
+import { ProjectOwnerGuard } from '../project/project.owner.guard';
 
 @Controller('setting/project')
 export class SettingProjectController {
   constructor(private readonly settingProjectService: SettingProjectService) {}
   @Patch('/add/:projectId')
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectOwnerGuard)
   async addAllowProject(
     @Body() allowDto: AddAllowDto,
     @Param('projectId') projectId: string,
     @Req() req: RequestWithUser,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.settingProjectService.addAllowProject(
       allowDto,
@@ -34,12 +40,15 @@ export class SettingProjectController {
   }
 
   @Patch('/update/:projectId')
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectOwnerGuard)
   async updateAllowProject(
     @Body() allowDto: AddAllowDto,
     @Param('projectId') projectId: string,
     @Req() req: RequestWithUser,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.settingProjectService.updateProjectAllow(
       allowDto,
@@ -49,12 +58,15 @@ export class SettingProjectController {
   }
 
   @Patch('/delete/:projectId')
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectOwnerGuard)
   async deleteAllowProject(
     @Body() allowDto: DeleteAllowDto,
     @Param('projectId') projectId: string,
     @Req() req: RequestWithUser,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.settingProjectService.deleteAllowProject(
       allowDto,
@@ -63,22 +75,28 @@ export class SettingProjectController {
     );
   }
   @Post('/discount/:projectId')
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectOwnerGuard)
   async addDiscount(
     @Body() dto: DiscountDto,
     @Param('projectId') projectId: string,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.settingProjectService.addDiscount(dto, objectId);
   }
 
   @Post('/lowEstimates/:projectId')
-  @UseGuards(ProjectGuard)
+  @UseGuards(ProjectOwnerGuard)
   async addLowEstimates(
     @Body() dto: DiscountDto,
     @Param('projectId') projectId: string,
     @Req() req: RequestWithUser,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.settingProjectService.addLowEstimates(dto, objectId, req);
   }
